Refactor cognito-auth & auth.service
